requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic[email]==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
python-dotenv==1.2.1
psycopg[binary]==3.3.2
sqlalchemy==2.0.45
alembic==1.18.0
polars==1.37.0
pandas==2.3.3
numpy==2.4.1
scikit-learn==1.8.0
scipy==1.17.0
joblib==1.5.3
pytz==2025.2
httpx==0.28.1
slowapi==0.1.9
cachetools==6.2.4
wheel==0.45.1
